Update: change file structure of stru_oper
