+++ enzy_htp/structure/structure_operation/__init__.py
@@ -4,3 +4,4 @@ Author: Qianzhen (QZ) Shao, <shaoqz@icloud.com>
 Date: 2022-09-19
 """
 from .general import *
+from .residue import *
